parsing with files and data from server works
simplified and standardized the way I handle the data I get. I have a way to parse single VTODO entries which is called each time it is needed. I also changed the tablewisget to not show all the fields available, but only the ones I need. That's because data from the server is going to be stored in the DOM object and not the table.